Fix for bug in recalculateParticipantMutation -> DeleteFinancialOverviewParticipantProject->delete() Checks in mutations not necessary for concept records.

+++ app/Helpers/Delete/Models/DeleteFinancialOverviewParticipantProject.php
@@ -46,13 +46,18 @@ class DeleteFinancialOverviewParticipantProject implements DeleteInterface
      */
     public function delete()
     {
-        $this->canDelete();
+        if (! $this->canDelete()) {
+            return $this->errorMessage;
+        }
+
         $this->deleteModels();
         $this->dissociateRelations();
         $this->deleteRelations();
         $this->customDeleteActions();
-        $this->financialOverviewParticipantProject->delete();
 
+        if (count($this->errorMessage) === 0) {
+            $this->financialOverviewParticipantProject->delete();
+        }
         return $this->errorMessage;
     }
 
@@ -60,21 +65,36 @@ class DeleteFinancialOverviewParticipantProject implements DeleteInterface
      */
     public function canDelete()
     {
+        $isDraft = $this->financialOverviewParticipantProject->status_id === 'concept';
+
+        if ($isDraft) {
+            return true;
+        }
+
+        $foDescription = $this->financialOverviewParticipantProject->financialOverview?->description ?? '*onbekend*';
+        $projectId = $this->financialOverviewParticipantProject?->financialOverviewProject?->project_id ?? '?';
+        $projectCode = $this->financialOverviewParticipantProject?->financialOverviewProject?->project?->code ?? 'onbekend';
+        $participationId = $this->financialOverviewParticipantProject?->participant_project_id ?? '?';
+        $contactId = $this->financialOverviewParticipantProject?->contact_id ?? '?';
+        $contactName = $this->financialOverviewParticipantProject?->contact?->full_name_fnf ?? '*contact onbekend*';
+
         if($this->financialOverviewParticipantProject->status_id === 'sent'){
-            array_push($this->errorMessage, "Er zijn al waardestaten voor deelnemer verzonden.");
+            array_push($this->errorMessage, "Waardestaat " . $foDescription . " voor deelnemer " . $contactName . " (" . $participationId . ") verzonden bij project " . $projectCode . " (" . $projectId . ")." );
         }
         $hasFinancialOverviewDefinitive = ParticipantMutation::where('participation_id', $this->financialOverviewParticipantProject->participant_project_id)
             ->where('financial_overview_definitive', true)->exists();
         if($hasFinancialOverviewDefinitive){
-            array_push($this->errorMessage, "Er zijn al mutaties voor deelnemer verwerkt in een definitieve project waarde staat.");
+            array_push($this->errorMessage, "Waardestaat " . $foDescription . " is al definitief voor deelnemer " . $contactName . " (" . $participationId . ") met mutaties bij project " . $projectCode . " (" . $projectId . ").");
         }
         $hasFinancialOverviewContactSent = FinancialOverviewContact::where('financial_overview_id',  $this->financialOverviewParticipantProject->financialOverviewProject->financial_overview_id)
             ->where('contact_id',  $this->financialOverviewParticipantProject->contact_id)
             ->where('status_id', 'sent')->exists();
 
         if($hasFinancialOverviewContactSent){
-            array_push($this->errorMessage, "Er zijn al waardestaten voor contact verzonden.");
+            array_push($this->errorMessage, "Waardestaat " . $foDescription . " voor contact " . $contactName . " (" . $contactId . ") verzonden bij project " . $projectCode . " (" . $projectId . ")." );
         }
+
+        return count($this->errorMessage) === 0;
     }
 
     /** Deletes models recursive
